Lots of optmizations
Move everything onto a non-default cuda stream. This mainly meant making
the int / size_t results of kernels that are used as arguments for
subsequent kernel launches into HostMemory async memcpys rather than
copying directly into a single int variable.  This way the copies can be
truly async.

Similary, moved the cuda memset() call into a separate stream so it runs
in parallel with device compute work.

Split up color->grey and input->decimated image processing.  Doing so
has a number of benefits. First off, the color->grey can be run on a
separate stream since it is independent of the rest of the cuda work
(the output grey image is used by CPU decoding towards the end of the
pipe).  It also helps thread utilization, since the combined process had
~25% utilization (most of the threads were idle while 1/4 were doing the
decimate).  Finally, the Mono8 case could be optimized because the input
is the same as the greyscale, so the entire color->grey / memcpy d2h
could be skipped in favor of a pointer assignment.

Initial work on making device code use 32 bit values rather than 64
wherever possible. This includes size_t -> uint32_t and double->float in
some cases. There's potentially more here to find but this commit
handled some of the low hanging fruit.